+++ protected/controllers/OrigenesController.php
@@ -112,7 +112,10 @@ class OrigenesController extends Controller
 		if(Yii::app()->request->isPostRequest)
 		{
 			// we only allow deletion via POST request
-			$this->loadModel($id)->delete();
+			//Falta mensaje emergente!!
+			$origenes = Guias::model()->find('id_origen=:cIDOrigen',array(':cIDOrigen'=>$id));
+			if (!(count($origenes)))
+				$this->loadModel($id)->delete();
 
 			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 			if(!isset($_GET['ajax']))
